+++ src/dev/lib/start.ts
@@ -14,5 +14,13 @@ export default function jestStart() {
       expect(kernel.getModCount()).toBe(modLen === null ? 2 : modLen);
       expect(kernel.getState()).toBe('running');
     });
+    test('kernel info', async () => {
+      expect(kernel.getModCount(true)).toBe(
+        modLen === null ? 2 + 2 : modLen + 2
+      );
+      expect(kernel.getServiceList(true).length).toBeGreaterThan(0);
+      expect(kernel.getActionList(true).length).toBeGreaterThan(0);
+      expect(kernel.getState()).toBe('running');
+    });
   });
 }